Move IndentingXMLStreamWriter from display to core/framework
.. and implement doSave for Data Browser
